+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 0.3.7
 Add `cutaway` methods to construct a vertical 2d slice of a 3d shape.
+Add `rotateAroundCenter` method to `PIXI.Rectangle`.
 
 ## 0.3.6
 Fix for tile pixel cache in v12.